refactor lang visitors
