+++ shards/lang/src/print.rs
@@ -1,13 +1,23 @@
-#[cfg(test)]
-use crate::read::process_program;
-#[cfg(test)]
+use crate::ast_visitor::*;
+use crate::ast::*;
+
 use pest::Parser;
 
-use crate::ast::*;
-#[cfg(test)]
-use crate::read::ReadEnv;
-
 const INDENT_LENGTH: usize = 2;
+
+struct Context {
+  indent: usize,
+  previous: Option<*const Context>,
+}
+
+impl Default for Context {
+  fn default() -> Self {
+    Context {
+      indent: 0,
+      previous: None,
+    }
+  }
+}
 
 impl Identifier {
   fn to_string(&self) -> String {
@@ -51,85 +61,234 @@ fn format_f32(f: f32) -> String {
   }
 }
 
-struct Context {
-  indent: usize,
-  previous: Option<*const Context>,
+pub struct PrintVisitor {
+  context: Context,
+  output: String,
 }
 
-impl Default for Context {
-  fn default() -> Self {
-    Context {
-      indent: 0,
-      previous: None,
+impl PrintVisitor {
+  pub fn new() -> Self {
+    PrintVisitor {
+      context: Context::default(),
+      output: String::new(),
     }
+  }
+
+  pub fn result(self) -> String {
+    self.output
+  }
+
+  fn indent(&mut self) {
+    self.context.indent += INDENT_LENGTH;
+  }
+
+  fn dedent(&mut self) {
+    if self.context.indent >= INDENT_LENGTH {
+      self.context.indent -= INDENT_LENGTH;
+    }
+  }
+
+  fn write(&mut self, text: &str) {
+    self.output.push_str(text);
+  }
+
+  fn write_line(&mut self, text: &str) {
+    self.write(&" ".repeat(self.context.indent));
+    self.write(text);
+    self.write("\n");
   }
 }
 
-impl Value {
-  fn to_string(&self, context: &mut Context) -> String {
-    match self {
-      Value::None => String::from("none"),
-      Value::Identifier(s) => s.to_string(),
-      Value::Boolean(b) => format!("{}", b),
-      Value::Enum(a, b) => format!("{}.{}", a, b),
-      Value::Number(n) => n.to_string(),
+impl AstVisitor for PrintVisitor {
+  fn visit_program(&mut self, program: &Program) {
+    program.metadata.accept(self);
+    program.sequence.accept(self);
+  }
+
+  fn visit_sequence(&mut self, sequence: &Sequence) {
+    for statement in &sequence.statements {
+      statement.accept(self);
+    }
+  }
+
+  fn visit_statement(&mut self, statement: &Statement) {
+    match statement {
+      Statement::Assignment(assignment) => assignment.accept(self),
+      Statement::Pipeline(pipeline) => pipeline.accept(self),
+    }
+    self.write("\n")
+  }
+
+  fn visit_assignment(&mut self, assignment: &Assignment) {
+    match assignment {
+      Assignment::AssignRef(pipeline, identifier) => {
+        pipeline.accept(self);
+        self.write(&format!(" = {}", identifier.to_string()));
+      }
+      Assignment::AssignSet(pipeline, identifier) => {
+        pipeline.accept(self);
+        self.write(&format!(" >= {}", identifier.to_string()));
+      }
+      Assignment::AssignUpd(pipeline, identifier) => {
+        pipeline.accept(self);
+        self.write(&format!(" > {}", identifier.to_string()));
+      }
+      Assignment::AssignPush(pipeline, identifier) => {
+        pipeline.accept(self);
+        self.write(&format!(" >> {}", identifier.to_string()));
+      }
+    }
+  }
+
+  fn visit_pipeline(&mut self, pipeline: &Pipeline) {
+    for (i, block) in pipeline.blocks.iter().enumerate() {
+      if i > 0 {
+        self.write(" | ");
+      }
+      block.accept(self);
+    }
+  }
+
+  fn visit_block(&mut self, block: &Block) {
+    match &block.content {
+      BlockContent::Empty => self.write(""),
+      BlockContent::Shard(function) => function.accept(self),
+      BlockContent::Func(function) => {
+        self.write("@");
+        function.accept(self);
+      }
+      BlockContent::Shards(sequence) => {
+        self.write("{");
+        sequence.accept(self);
+        self.write("}");
+      }
+      BlockContent::Const(value) => value.accept(self),
+      BlockContent::TakeTable(identifier, path) => {
+        let path = path.iter().map(|p| p.to_string()).collect::<Vec<String>>();
+        self.write(&format!("{}:{}", identifier.to_string(), path.join(":")));
+      }
+      BlockContent::TakeSeq(identifier, path) => {
+        self.write(&format!(
+          "{}:{}",
+          identifier.to_string(),
+          path
+            .iter()
+            .map(|p| p.to_string())
+            .collect::<Vec<String>>()
+            .join(":")
+        ));
+      }
+      BlockContent::EvalExpr(sequence) => {
+        self.write("#(");
+        sequence.accept(self);
+        self.write(")");
+      }
+      BlockContent::Expr(sequence) => {
+        self.write("(");
+        sequence.accept(self);
+        self.write(")");
+      }
+      BlockContent::Program(program) => {
+        program.sequence.accept(self);
+      }
+    }
+  }
+
+  fn visit_function(&mut self, function: &Function) {
+    self.write(&function.name.to_string());
+    if let Some(params) = &function.params {
+      let params_str = params
+        .iter()
+        .map(|v| {
+          let mut visitor = PrintVisitor::new();
+          v.accept(&mut visitor);
+          visitor.result()
+        })
+        .collect::<Vec<String>>()
+        .join(" ");
+      if !params_str.is_empty() {
+        self.write(&format!("({})", params_str));
+      }
+    }
+  }
+
+  fn visit_param(&mut self, param: &Param) {
+    if let Some(name) = &param.name {
+      self.write(&format!("{}: ", name));
+    }
+    param.value.accept(self);
+  }
+
+  fn visit_identifier(&mut self, identifier: &Identifier) {
+    self.write(&identifier.to_string());
+  }
+
+  fn visit_value(&mut self, value: &Value) {
+    match value {
+      Value::None => self.write("none"),
+      Value::Identifier(identifier) => self.visit_identifier(identifier),
+      Value::Boolean(b) => self.write(&format!("{}", b)),
+      Value::Enum(a, b) => self.write(&format!("{}::{}", a, b)),
+      Value::Number(number) => self.write(&number.to_string()),
       Value::String(s) => {
-        // if within the string there are quotes, we need to escape them
         let escaped = s.replace("\"", "\\\"");
-        format!("\"{}\"", escaped)
+        self.write(&format!("\"{}\"", escaped));
       }
       Value::Bytes(b) => {
         let bytes_slice = b.0.as_ref();
-        // build a big hex string from the bytes
-        let mut hex_string = String::new();
-        for byte in bytes_slice {
-          hex_string.push_str(&format!("{:02x}", byte));
-        }
-        format!("@bytes(0x{})", hex_string)
+        let hex_string: String = bytes_slice
+          .iter()
+          .map(|byte| format!("{:02x}", byte))
+          .collect();
+        self.write(&format!("@bytes(0x{})", hex_string));
       }
-      Value::Int2(arr) => format!("({} {})", arr[0], arr[1]),
-      Value::Int3(arr) => format!("({} {} {})", arr[0], arr[1], arr[2]),
-      Value::Int4(arr) => format!("({} {} {} {})", arr[0], arr[1], arr[2], arr[3]),
-      Value::Float2(arr) => format!("({} {})", format_f64(arr[0]), format_f64(arr[1])),
-      Value::Float3(arr) => format!(
+      Value::Int2(arr) => self.write(&format!("({} {})", arr[0], arr[1])),
+      Value::Int3(arr) => self.write(&format!("({} {} {})", arr[0], arr[1], arr[2])),
+      Value::Int4(arr) => self.write(&format!("({} {} {} {})", arr[0], arr[1], arr[2], arr[3])),
+      Value::Float2(arr) => self.write(&format!("({} {})", format_f64(arr[0]), format_f64(arr[1]))),
+      Value::Float3(arr) => self.write(&format!(
         "({} {} {})",
         format_f32(arr[0]),
         format_f32(arr[1]),
         format_f32(arr[2])
-      ),
-      Value::Float4(arr) => format!(
+      )),
+      Value::Float4(arr) => self.write(&format!(
         "({} {} {} {})",
         format_f32(arr[0]),
         format_f32(arr[1]),
         format_f32(arr[2]),
         format_f32(arr[3])
-      ),
+      )),
       Value::Seq(values) => {
         let values_str = values
           .iter()
-          .map(|v| v.to_string(context))
+          .map(|v| {
+            let mut visitor = PrintVisitor::new();
+            v.accept(&mut visitor);
+            visitor.result()
+          })
           .collect::<Vec<String>>()
           .join(" ");
-        format!("[{}]", values_str)
+        self.write(&format!("[{}]", values_str));
       }
-      Value::TakeTable(name, path) => {
+      Value::TakeTable(identifier, path) => {
         let path = path.iter().map(|p| p.to_string()).collect::<Vec<String>>();
         let path = path.join(":");
-        format!("{}:{}", name.to_string(), path)
+        self.write(&format!("{}:{}", identifier.to_string(), path));
       }
-      Value::TakeSeq(name, path) => {
+      Value::TakeSeq(identifier, path) => {
         let path_str = path
           .iter()
           .map(|p| p.to_string())
           .collect::<Vec<String>>()
           .join(":");
-        format!("{}:{}", name.to_string(), path_str)
+        self.write(&format!("{}:{}", identifier.to_string(), path_str));
       }
-      Value::Int8(arr) => format!(
+      Value::Int8(arr) => self.write(&format!(
         "({} {} {} {} {} {} {} {})",
         arr[0], arr[1], arr[2], arr[3], arr[4], arr[5], arr[6], arr[7]
-      ),
-      Value::Int16(arr) => format!(
+      )),
+      Value::Int16(arr) => self.write(&format!(
         "({} {} {} {} {} {} {} {} {} {} {} {} {} {} {} {})",
         arr[0],
         arr[1],
@@ -147,169 +306,71 @@ impl Value {
         arr[13],
         arr[14],
         arr[15]
-      ),
+      )),
       Value::Table(values) => {
         let values_str = values
           .iter()
-          .map(|(k, v)| format!("{}: {}", k.to_string(context), v.to_string(context)))
+          .map(|(k, v)| {
+            let mut key_visitor = PrintVisitor::new();
+            k.accept(&mut key_visitor);
+            let key_str = key_visitor.result();
+
+            let mut value_visitor = PrintVisitor::new();
+            v.accept(&mut value_visitor);
+            let value_str = value_visitor.result();
+
+            format!("{}: {}", key_str, value_str)
+          })
           .collect::<Vec<String>>()
           .join(" ");
-        format!("{{{}}}", values_str)
+        self.write(&format!("{{{}}}", values_str));
       }
-      Value::Shards(seq) => format!("{{{}}}", seq.to_string(context)),
-      Value::Shard(func) => format!("{}", func.to_string(context)),
-      Value::EvalExpr(seq) => format!("#({})", seq.to_string(context)),
-      Value::Expr(seq) => format!("({})", seq.to_string(context)),
-      Value::Func(func) => format!("@{}", func.to_string(context)),
-    }
-  }
-}
-
-impl Param {
-  fn to_string(&self, context: &mut Context) -> String {
-    match &self.name {
-      Some(name) => format!("{}: {}", name, self.value.to_string(context)),
-      None => self.value.to_string(context),
-    }
-  }
-}
-
-impl Function {
-  fn to_string(&self, context: &mut Context) -> String {
-    let params: Vec<String> = match &self.params {
-      Some(params) => params.iter().map(|p| p.to_string(context)).collect(),
-      None => vec![],
-    };
-    if params.is_empty() {
-      self.name.to_string()
-    } else {
-      format!("{}({})", self.name.to_string(), params.join(" "))
-    }
-  }
-}
-
-impl BlockContent {
-  fn to_string(&self, context: &mut Context) -> String {
-    match self {
-      BlockContent::Shard(func) => func.to_string(context),
-      BlockContent::Func(func) => format!("@{}", func.to_string(context)),
-      BlockContent::Shards(seq) => format!("{{{}}}", seq.to_string(context)),
-      BlockContent::Const(value) => value.to_string(context),
-      BlockContent::TakeTable(name, path) => {
-        let path = path.iter().map(|p| p.to_string()).collect::<Vec<String>>();
-        let path = path.join(":");
-        format!("{}:{}", name.to_string(), path)
+      Value::Shards(seq) => {
+        self.write("{");
+        seq.accept(self);
+        self.write("}");
       }
-      BlockContent::TakeSeq(name, path) => {
-        let path_str = path
-          .iter()
-          .map(|p| p.to_string())
-          .collect::<Vec<String>>()
-          .join(":");
-        format!("{}:{}", name.to_string(), path_str)
+      Value::Shard(func) => func.accept(self),
+      Value::EvalExpr(seq) => {
+        self.write("#(");
+        seq.accept(self);
+        self.write(")");
       }
-      BlockContent::EvalExpr(seq) => format!("#({})", seq.to_string(context)),
-      BlockContent::Expr(seq) => format!("({})", seq.to_string(context)),
-      BlockContent::Empty => "".to_string(),
-      BlockContent::Program(p) => p.sequence.to_string(context),
-    }
-  }
-}
-
-impl Block {
-  fn to_string(&self, context: &mut Context) -> String {
-    self.content.to_string(context)
-  }
-}
-
-impl Pipeline {
-  fn to_string(&self, context: &mut Context) -> String {
-    let blocks_str = self
-      .blocks
-      .iter()
-      .map(|b| b.to_string(context))
-      .collect::<Vec<String>>()
-      .join(" | ");
-    format!("{}", blocks_str)
-  }
-}
-
-impl Assignment {
-  fn to_string(&self, context: &mut Context) -> String {
-    match self {
-      Assignment::AssignRef(pipeline, name) => {
-        format!("{} = {}", pipeline.to_string(context), name.to_string())
+      Value::Expr(seq) => {
+        self.write("(");
+        seq.accept(self);
+        self.write(")");
       }
-      Assignment::AssignSet(pipeline, name) => {
-        format!("{} >= {}", pipeline.to_string(context), name.to_string())
-      }
-      Assignment::AssignUpd(pipeline, name) => {
-        format!("{} > {}", pipeline.to_string(context), name.to_string())
-      }
-      Assignment::AssignPush(pipeline, name) => {
-        format!("{} >> {}", pipeline.to_string(context), name.to_string())
+      Value::Func(func) => {
+        self.write("@");
+        func.accept(self);
       }
     }
   }
-}
 
-impl Statement {
-  fn to_string(&self, context: &mut Context) -> String {
-    match self {
-      Statement::Assignment(assign) => assign.to_string(context),
-      Statement::Pipeline(pipeline) => pipeline.to_string(context),
-    }
-  }
-}
-
-impl Sequence {
-  fn to_string(&self, context: &mut Context) -> String {
-    let mut inner_context = Context::default();
-    inner_context.previous = Some(context);
-    if self.statements.len() > 1 && context.previous.is_some() {
-      inner_context.indent = context.indent + INDENT_LENGTH;
-    }
-
-    let statements_str = self
-      .statements
-      .iter()
-      .map(|s| s.to_string(&mut inner_context))
-      .collect::<Vec<String>>()
-      .join("\n");
-
-    let indent_str = " ".repeat(inner_context.indent);
-    let new_line = if inner_context.indent > 0 { "\n" } else { "" };
-    let indented_blocks_str = statements_str.replace("\n", &format!("\n{}", indent_str));
-    format!(
-      "{}{}{}{}",
-      new_line, indent_str, indented_blocks_str, new_line
-    )
+  fn visit_metadata(&mut self, metadata: &Metadata) {
+    self.write(&metadata.name.to_string());
   }
 }
 
 pub fn print_ast(ast: &Sequence) -> String {
-  let mut context = Context::default();
-  ast.to_string(&mut context)
+  let mut visitor = PrintVisitor::new();
+  ast.accept(&mut visitor);
+  visitor.result()
 }
 
 #[test]
-fn test_print1() {
+fn test1() {
+  // use std::num::NonZeroUsize;
+  // pest::set_call_limit(NonZeroUsize::new(25000));
+  // let code = include_str!("nested.shs");
   let code = include_str!("sample1.shs");
-  let successful_parse = ShardsParser::parse(Rule::Program, code).unwrap();
-  let mut env = ReadEnv::new("", ".", ".");
-  let seq = process_program(successful_parse.into_iter().next().unwrap(), &mut env).unwrap();
+  let successful_parse = crate::ShardsParser::parse(Rule::Program, code).unwrap();
+  let mut env = crate::read::ReadEnv::new("", ".", ".");
+  let seq = crate::read::process_program(successful_parse.into_iter().next().unwrap(), &mut env).unwrap();
   let seq = seq.sequence;
-  let s = print_ast(&seq);
-  println!("{}", s);
-}
 
-#[test]
-fn test_print2() {
-  let code = include_str!("explained.shs");
-  let successful_parse = ShardsParser::parse(Rule::Program, code).unwrap();
-  let mut env = ReadEnv::new("", ".", ".");
-  let seq = process_program(successful_parse.into_iter().next().unwrap(), &mut env).unwrap();
-  let seq = seq.sequence;
-  let s = print_ast(&seq);
-  println!("{}", s);
+  let printed = print_ast(&seq);
+
+  println!("{}", printed);
 }
